requirements.txt
Django==5.1.7
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pillow==11.1.0
crispy-bootstrap5==2025.4
django-filter==25.1
Markdown==3.8
djangorestframework==3.16.0
